[DEV-0] Fixed stats

+++ Assets/Scripts/Common/Reactive/ReactiveValueUpdatedEventArgs.cs
@@ -2,11 +2,11 @@
 {
     public struct ReactiveValueUpdatedEventArgs
     {
-        public readonly float after;
-        public readonly float before;
-        public readonly float max;
+        public readonly int after;
+        public readonly int before;
+        public readonly int max;
 
-        public ReactiveValueUpdatedEventArgs(float before, float after, float max)
+        public ReactiveValueUpdatedEventArgs(int before, int after, int max)
         {
             this.before = before;
             this.after = after;
